Added Identity Service

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -9,6 +9,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -63,12 +64,13 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
             Auction auction = _mapper.Map<Auction>(auctionDto);
             // ToDo: Add current user as seller
-            auction.Seller = "test";
+            auction.Seller = User.Identity.Name;
 
             _context.Auctions.Add(auction);
 
@@ -86,6 +88,7 @@ namespace AuctionService.Controllers
                 new {auction.Id}, _mapper.Map<AuctionDto>(auction));
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
@@ -97,7 +100,7 @@ namespace AuctionService.Controllers
                 return NotFound();
             }
 
-            // ToDo: Check for user authentication for update operation
+            if(auction.Seller != User.Identity.Name) return Forbid();
 
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -117,6 +120,7 @@ namespace AuctionService.Controllers
             return BadRequest("Could not update the changes!");
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
@@ -127,7 +131,8 @@ namespace AuctionService.Controllers
                 return NotFound();
             }
 
-            // ToDo: Check for user authentication for Delete operation
+            if(auction.Seller != User.Identity.Name) return Forbid();
+            
             _context.Auctions.Remove(auction);
 
             await _publishEndpoint.Publish<AuctionDeleted>(new {Id = auction.Id.ToString()});
